backend/requirements.txt
# Railway fallback requirements file
# Primary dependency management is via uv and pyproject.toml
# This file is for Railway's auto-detection

# Core dependencies
a2a-sdk>=0.3.0
a2a-server>=0.6.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
google-adk>=1.18.0
web3>=6.15.0
ag-ui-adk>=0.0.1
# hiero-sdk-python>=0.1.0  # Optional - install manually if needed (has dependency conflicts)

